+++ src/Fabulous.MauiControls/Views/Controls/TimePicker.fs
@@ -85,7 +85,7 @@ type TimePickerModifiers =
     static member inline font
         (
             this: WidgetBuilder<'msg, #ITimePicker>,
-            ?size: double,
+            ?size: float,
             ?namedSize: NamedSize,
             ?attributes: FontAttributes,
             ?fontFamily: string
